create bookmark file and move bookmark struct and methods into here
